+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 1.0.1
+
+- Fix display of active item when external code is hidden
+
+## 1.0.0
+
+- Display currently active stack item
+- Open a source file when clicking in a node
+
 ## 0.2.1
 
 - Fix hover tooltip when external code is hidden